Disable caching of thumbnail and poster images
Thumbnail and poster images can be updated by collection staff and with cache enabled it appeared that they weren't changing due to the 30 minute cache expiry.  Disabling the cache resolves the UX bug but might cause a slightly heavier load on rails and solr as a result.  A different approach to cache expiration is needed that is change-aware and not time based.