Fix broken 'hls' service.
The image which was used by the 'hls' container,
'docker.io/avalonmediasystem/nginx:ecs', is built separately from this
repository and built using the build context from the 'aws_ecs' branch.
The script that processes the Nginx configuration template at startup
inside the 'hls' container differs between the 'aws_ecs' and 'aws_min'
branches: the 'aws_ecs' version processes the 'AUTH_URL' and
'PROXY_PASS_URL' variables; whereas the 'aws_min' version processes the
'AVALON_DOMAIN' and 'AVALON_STREAMING_BUCKET' variables.

With f9576a4, this bind-mounted the Nginx configuration template into
the 'hls' container; previously, the 'hls' container used the template
that was baked into the image from the 'aws_ecs' branch, and it matched
the template-processing script. Because the Nginx configuration template
in this branch references the 'AVALON_STREAMING_BUCKET' and
'AVALON_DOMAIN' variables but the script didn't substitute either, Nginx
aborted when it tried and failed to process those variables itself.

This changes the 'hls' container to use the 'aws_min' version of the
image, to provide it with the environment variables referenced by the
Nginx configuration template in this branch, and to remove the unused
's3-helper' container.